added more comments and more code splitting
