+++ provider/pkg/provider/local/baseOutputs.go
