+++ provider/pkg/provider/local/commandController.go
@@ -15,23 +15,9 @@
 package local
 
 import (
-	"bytes"
-	"fmt"
-	"io"
-	"io/fs"
-	"os"
-	"os/exec"
-	"path"
-	"runtime"
-	"strings"
-
-	"github.com/gobwas/glob"
 	p "github.com/pulumi/pulumi-go-provider"
 	"github.com/pulumi/pulumi-go-provider/infer"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/diag"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource"
-
-	"github.com/pulumi/pulumi-command/provider/pkg/provider/util"
 )
 
 // The following statements are not required. They are type assertions to indicate to Go that Command implements the following interfaces.
@@ -43,6 +29,31 @@ var _ = (infer.CustomUpdate[CommandInputs, CommandOutputs])((*Command)(nil))
 var _ = (infer.CustomDelete[CommandOutputs])((*Command)(nil))
 var _ = (infer.ExplicitDependencies[CommandInputs, CommandOutputs])((*Command)(nil))
 
+// This is the Create method. This will be run on every Command resource creation.
+func (c *Command) Create(ctx p.Context, name string, input CommandInputs, preview bool) (string, CommandOutputs, error) {
+	state := CommandOutputs{CommandInputs: input}
+	id, err := resource.NewUniqueHex(name, 8, 0)
+	if err != nil {
+		return id, state, err
+	}
+
+	// If in preview, don't run the command.
+	if preview {
+		return id, state, nil
+	}
+	if input.Create == nil {
+		return id, state, nil
+	}
+	cmd := *input.Create
+	state.Stdout, state.Stderr, err = state.run(ctx, cmd)
+	return id, state, err
+}
+
+// WireDependencies is relevant to secrets handling. This method indicates the what Inputs
+// the Outputs are derived from. If an output is derived from a secret input, the output
+// will be a secret.
+
+// This naive implementation conveys that every output is derived from all inputs.
 func (r *Command) WireDependencies(f infer.FieldSelector, args *CommandInputs, state *CommandOutputs) {
 
 	// get BaseInputs
@@ -61,7 +72,7 @@ func (r *Command) WireDependencies(f infer.FieldSelector, args *CommandInputs, s
 	f.OutputField(&state.AssetPaths).DependsOn(assetPathsInput)
 	f.OutputField(&state.ArchivePaths).DependsOn(archivePathsInput)
 
-	// set
+	// get CommandInputs
 	triggersInput := f.InputField(&args.Triggers)
 	createInput := f.InputField(&args.Create)
 	updateInput := f.InputField(&args.Update)
@@ -73,6 +84,7 @@ func (r *Command) WireDependencies(f infer.FieldSelector, args *CommandInputs, s
 	f.OutputField(&state.Update).DependsOn(updateInput)
 	f.OutputField(&state.Delete).DependsOn(deleteInput)
 
+	// Mark Stdout as derived from all inputs
 	f.OutputField(&state.Stdout).DependsOn(
 		triggersInput,
 		createInput,
@@ -85,6 +97,8 @@ func (r *Command) WireDependencies(f infer.FieldSelector, args *CommandInputs, s
 		assetPathsInput,
 		archivePathsInput,
 	)
+
+	// Mark Stderr as derived from all inputs
 	f.OutputField(&state.Stderr).DependsOn(
 		triggersInput,
 		createInput,
@@ -99,185 +113,32 @@ func (r *Command) WireDependencies(f infer.FieldSelector, args *CommandInputs, s
 	)
 }
 
-func (c *Command) Create(ctx p.Context, name string, input CommandInputs, preview bool) (string, CommandOutputs, error) {
-	state := CommandOutputs{CommandInputs: input}
-	id, err := resource.NewUniqueHex(name, 8, 0)
-	if err != nil {
-		return id, state, err
-	}
-
-	if preview {
-		return id, state, nil
-	}
-	if input.Create == nil {
-		return id, state, nil
-	}
-	cmd := *input.Create
-	state.Stdout, state.Stderr, err = state.run(ctx, cmd)
-	return id, state, err
-}
-
+// The Update method will be run on every update.
 func (c *Command) Update(ctx p.Context, id string, olds CommandOutputs, news CommandInputs, preview bool) (CommandOutputs, error) {
 	state := CommandOutputs{CommandInputs: news}
+	// If in preview, don't run the command.
 	if preview {
 		return state, nil
 	}
-	var err error
+	// Use Create command if Update is unspecified.
+	cmd := news.Create
 	if news.Update != nil {
-		state.Stdout, state.Stderr, err = state.run(ctx, *news.Update)
-	} else if news.Create != nil {
-		state.Stdout, state.Stderr, err = state.run(ctx, *news.Create)
+		cmd = news.Update
 	}
+	// If neither are specified, do nothing.
+	if cmd == nil {
+		return state, nil
+	}
+	var err error
+	state.Stdout, state.Stderr, err = state.run(ctx, *cmd)
 	return state, err
 }
 
+// The Delete method will run when the resource is deleted.
 func (c *Command) Delete(ctx p.Context, id string, props CommandOutputs) error {
 	if props.Delete == nil {
 		return nil
 	}
 	_, _, err := props.run(ctx, *props.Delete)
 	return err
-}
-
-func (c *CommandOutputs) run(ctx p.Context, command string) (string, string, error) {
-	var args []string
-	if c.Interpreter != nil && len(*c.Interpreter) > 0 {
-		args = append(args, *c.Interpreter...)
-	} else {
-		if runtime.GOOS == "windows" {
-			args = []string{"cmd", "/C"}
-		} else {
-			args = []string{"/bin/sh", "-c"}
-		}
-	}
-	args = append(args, command)
-
-	stdoutr, stdoutw, err := os.Pipe()
-	if err != nil {
-		return "", "", err
-	}
-	stderrr, stderrw, err := os.Pipe()
-	if err != nil {
-		return "", "", err
-	}
-
-	cmd := exec.CommandContext(ctx, args[0], args[1:]...)
-	cmd.Stdout = stdoutw
-	cmd.Stderr = stderrw
-	if c.Dir != nil {
-		cmd.Dir = *c.Dir
-	} else {
-		cmd.Dir, err = os.Getwd()
-		if err != nil {
-			return "", "", err
-		}
-	}
-	cmd.Env = os.Environ()
-	if c.Environment != nil {
-		for k, v := range *c.Environment {
-			cmd.Env = append(cmd.Env, fmt.Sprintf("%s=%s", k, v))
-		}
-	}
-
-	if c.Stdin != nil && len(*c.Stdin) > 0 {
-		cmd.Stdin = strings.NewReader(*c.Stdin)
-	}
-
-	var stdoutbuf bytes.Buffer
-	var stderrbuf bytes.Buffer
-
-	stdouttee := io.TeeReader(stdoutr, &stdoutbuf)
-	stderrtee := io.TeeReader(stderrr, &stderrbuf)
-
-	stdoutch := make(chan struct{})
-	stderrch := make(chan struct{})
-	go util.CopyOutput(ctx, stdouttee, stdoutch, diag.Debug)
-	go util.CopyOutput(ctx, stderrtee, stderrch, diag.Error)
-
-	err = cmd.Start()
-	if err == nil {
-		err = cmd.Wait()
-	}
-
-	stdoutw.Close()
-	stderrw.Close()
-
-	<-stdoutch
-	<-stderrch
-
-	if err != nil {
-		return "", "", err
-	}
-
-	if c.AssetPaths != nil {
-		assets, err := globAssets(cmd.Dir, *c.AssetPaths)
-		if err != nil {
-			return "", "", err
-		}
-		c.Assets = &assets
-	}
-
-	if c.ArchivePaths != nil {
-		archiveAssets := map[string]interface{}{}
-		assets, err := globAssets(cmd.Dir, *c.ArchivePaths)
-		if err != nil {
-			return "", "", err
-		}
-
-		for path, asset := range assets {
-			archiveAssets[path] = asset
-		}
-
-		archive, err := resource.NewAssetArchive(archiveAssets)
-		if err != nil {
-			return "", "", err
-		}
-		c.Archive = archive
-	}
-
-	return strings.TrimSuffix(stdoutbuf.String(), "\n"), strings.TrimSuffix(stderrbuf.String(), "\n"), nil
-}
-
-func globAssets(dir string, globs []string) (map[string]*resource.Asset, error) {
-	assets := map[string]*resource.Asset{}
-	compiledGlobs := make([]glob.Glob, len(globs))
-	isGlobExclude := make([]bool, len(globs))
-	for i, g := range globs {
-		isExclude := strings.HasPrefix(g, "!")
-		g = strings.TrimPrefix(g, "!")
-		compiled, err := glob.Compile(g, '/')
-		if err != nil {
-			return nil, err
-		}
-		compiledGlobs[i] = compiled
-		isGlobExclude[i] = isExclude
-	}
-
-	err := fs.WalkDir(os.DirFS(dir), ".", func(p string, d fs.DirEntry, err error) error {
-		if err != nil {
-			return err
-		}
-		if d.IsDir() {
-			return nil
-		}
-		for i, g := range compiledGlobs {
-			matched := g.Match(p)
-			if !matched {
-				continue
-			}
-			if isGlobExclude[i] {
-				delete(assets, p)
-			} else {
-				assets[p], err = resource.NewPathAsset(path.Join(dir, p))
-				if err != nil {
-					return err
-				}
-			}
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	return assets, nil
 }
