+++ provider/pkg/provider/remote/copyfile.go
@@ -22,7 +22,7 @@ type CopyFile struct{}
 
 var _ = (infer.Annotated)((*CopyFile)(nil))
 
-// CopyFile implements Annotate which allows you to attach descriptions to the CopyFile resource and its fields.
+// CopyFile implements Annotate which allows you to attach descriptions to the CopyFile resource.
 func (c *CopyFile) Annotate(a infer.Annotator) {
 	a.Describe(&c, "Copy a local file to a remote host.")
 }
@@ -34,6 +34,7 @@ type CopyFileInputs struct {
 	RemotePath string         `pulumi:"remotePath"`
 }
 
+// CopyFile implements Annotate which allows you to attach descriptions to the CopyFile resource's fields.
 func (c *CopyFileInputs) Annotate(a infer.Annotator) {
 	a.Describe(&c.Connection, "The parameters with which to connect to the remote host.")
 	a.Describe(&c.Triggers, "Trigger replacements on changes to this input.")
